Serveur/requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
pydantic==1.9.0
python-dotenv==0.20.0
requests==2.27.1